refactor: Move MainActivity code to SmishingActivity (#2)
